### Description:
_<Describe proposed changes here...>_

### PR Checklist:
- [ ] PR title prefix added. (ICM, S360, Hot-Fix, Cherry-pick etc.)
- [ ] Dependent changes addressed. (Geneva config, Service Basics, Partner check-ins etc.)
- [ ] Integration tests added/updated.

### Testing:
- [ ] Manual E2E Testing. (Local, DF, Dev, CI, Perf, X-Region, X-Tenant)
- [ ] Automated E2E Testing. (Dev, UX)

## Type of change
- [ ] Bug fix _(non-breaking change which fixes an issue)_
- [ ] New feature _(non-breaking change which adds functionality)_
- [ ] Breaking change _(fix or feature that would cause existing functionality to be altered or to not work as expected)_
- [ ] Infrastructure/Configuration

### Additional Details (optional)
_<Add more details as necessary to help reviewers better understand the changes...>_

Add Tables to Domain model for self serve analytics

----
#### AI description  (iteration 1)
#### PR Classification
New feature

#### PR Summary
This pull request adds new tables to the domain model to support self-serve analytics, addressing the OKR addition to semantic models.
- `setup.sql`: Added external table creation scripts for CDE, CDEColumnAssignment, CDEDataProductAssignment, CDEGlossaryTermAssignment, Action, OKR, and KeyResult.
- `setup-rollback.sql`: Added scripts to drop the newly created external tables.
- `CatalogSparkJobComponent.cs`: Removed the `CosmosDBLinkedServiceName` argument from the Spark job configuration.

Related work items: #3641450

+++ src/DataDomain/DomainModel/src/main/scala/com/microsoft/azurepurview/dataestatehealth/domainmodel/okr/KeyResult.scala
@@ -4,7 +4,7 @@ import com.microsoft.azurepurview.dataestatehealth.domainmodel.common.Validator
 import org.apache.log4j.Logger
 import org.apache.spark.sql.expressions.Window
 import org.apache.spark.sql.functions.{col, row_number, when}
-import org.apache.spark.sql.types.{LongType, TimestampType}
+import org.apache.spark.sql.types.{IntegerType, LongType, TimestampType}
 import org.apache.spark.sql.{DataFrame, SparkSession}
 
 class KeyResult(spark: SparkSession, logger: Logger) {
@@ -90,9 +90,9 @@ class KeyResult(spark: SparkSession, logger: Logger) {
         .select(col("KeyResultId")
           , col("KeyResultDefintion")
           , col("Status")
-          , col("Progress")
-          , col("Goal")
-          , col("Max")
+          , col("Progress").cast(IntegerType)
+          , col("Goal").cast(IntegerType)
+          , col("Max").cast(IntegerType)
           , col("OKRId")
           , col("AccountId")
           , col("CreatedAt").alias("CreatedDatetime").cast(TimestampType)
